pyproject.toml
[tool.poetry]
name = "trainwithkaffka"
version = "0.1.0"
description = ""
authors = ["Tamagil <tamagiltamagil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
